Completed refactor; everything works
