+++ .gitignore
@@ -137,3 +137,9 @@ dist
 
 # macOS stuff
 .DS_Store
+
+# Project Specific Ignores
+frontend/node_modules/
+backend/node_modules/
+backend/srv.env
+backend/threat_intel.db
