--- api/srv.env
@@ -1,31 +1,0 @@
-# Server Configuration
-PORT=5050                          # Port for the backend server
-FRONTEND_URL=http://localhost:5173  # Frontend URL for CORS
-
-# API Keys
-SHODAN_API_KEY=HlvDD7cbdhneBv7WzGRZtZ1JSnCHQIP8
-HIBP_API_KEY=your_hibp_api_key
-
-# SQLite Database
-DB_PATH=./db/threat_intel.db
-
-# Email Configuration (for sending alerts)
-EMAIL_HOST=smtp.yourmailserver.com
-EMAIL_PORT=587
-EMAIL_USER=hmnilsen154@gmail.com
-EMAIL_PASS="rmfc supw mxvh lfse"
-ALERT_EMAIL_RECIPIENT=hmnilsen154@gmail.com  # ✅ ADD THIS LINE
-
-# Miscellaneous Configuration
-ALERT_WEBHOOK_URL=https://your-webhook-url.com
-LOGGING_LEVEL=debug
-
-# External Services Configuration
-OSINT_API_KEY=your_osint_api_key
-RISK_ANALYSIS_API_KEY=your_risk_analysis_api_key
-
-# Custom Configuration
-TEST_EMAIL=your_test_email@example.com
-
-# OpenAI
-OPENAI_API_KEY=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
