+++ backend/package.json
@@ -2,10 +2,10 @@
   "name": "threat-intel-api",
   "version": "1.0.0",
   "description": "Cybersecurity backend server with threat analysis, alerts, and OSINT integration.",
-  "main": "server.js",
+  "main": "src/server.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
+    "start": "node src/server.js",
+    "dev": "nodemon src/server.js"
   },
   "keywords": [
     "cybersecurity",
