+++ backend/src/blue_team_defense.js
