+++ backend/src/mitigation_recommendations.js
