+++ backend/src/threat_mitigation.js
