+++ frontend/src/components/ThreatDashboard.tsx
